updated JS code to use new Lambda backend for functionality.
